feat: add year option to /bws
